+++ src/SWP/Bundle/CoreBundle/SWPCoreBundle.php
@@ -14,6 +14,8 @@
 
 namespace SWP\Bundle\CoreBundle;
 
+use SWP\Bundle\CoreBundle\DependencyInjection\Compiler\AddCustomTwigCachePass;
+use SWP\Bundle\CoreBundle\DependencyInjection\Compiler\OverrideArticleBodyProcessorPass;
 use SWP\Bundle\CoreBundle\DependencyInjection\Compiler\OverrideContainerProviderPass;
 use SWP\Bundle\CoreBundle\DependencyInjection\Compiler\OverrideDynamicRouterPass;
 use SWP\Bundle\CoreBundle\DependencyInjection\Compiler\OverrideMediaManagerPass;
@@ -50,6 +52,8 @@ class SWPCoreBundle extends Bundle
         $container->addCompilerPass(new OverrideTemplatesSystemPass());
         $container->addCompilerPass(new OverrideThemeLoaderPass());
         $container->addCompilerPass(new OverrideSettingsManagerPass());
+        $container->addCompilerPass(new OverrideArticleBodyProcessorPass());
+        $container->addCompilerPass(new AddCustomTwigCachePass());
     }
 
     /**