+++ src/SWP/Bundle/CoreBundle/Tests/Controller/RegistrationControllerTest.php
@@ -73,4 +73,27 @@ class RegistrationControllerTest extends WebTestCase
         self::assertTrue($client->getResponse()->isSuccessful());
         self::assertContains('The user has been created successfully.', $client->getResponse()->getContent());
     }
+
+    public function testValidation()
+    {
+        $client = static::createClient();
+        $client->enableProfiler();
+        $client->request('POST', $this->router->generate('swp_api_core_register_user'), [
+            'user_registration' => [
+                'email' => 'notemail',
+                'username' => '',
+                'plainPassword' => [
+                    'first' => 'testPass',
+                    'second' => 'testPasss',
+                ],
+            ],
+        ]);
+
+        self::assertEquals(400, $client->getResponse()->getStatusCode());
+        $content = json_decode($client->getResponse()->getContent(), true);
+        self::assertArrayHasKey('errors', $content);
+        self::assertEquals('The email is not valid.', $content['errors']['children']['email']['errors'][0]);
+        self::assertEquals('Please enter a username.', $content['errors']['children']['username']['errors'][0]);
+        self::assertEquals('The entered passwords don\'t match.', $content['errors']['children']['plainPassword']['children']['first']['errors'][0]);
+    }
 }