+++ src/SWP/Bundle/CoreBundle/spec/Form/Type/TenantTypeSpec.php
@@ -57,10 +57,9 @@ class TenantTypeSpec extends ObjectBehavior
 
         $builder
             ->add('subdomain', TextType::class, [
-                'required' => true,
+                'required' => false,
                 'description' => 'Tenant subdomain',
                 'constraints' => [
-                    new NotBlank(),
                     new Length(['min' => 3]),
                 ],
             ])
@@ -69,9 +68,10 @@ class TenantTypeSpec extends ObjectBehavior
 
         $builder
             ->add('domainName', TextType::class, [
-                'required' => false,
+                'required' => true,
                 'description' => 'Tenant domain name',
                 'constraints' => [
+                    new NotBlank(),
                     new Length(['min' => 3]),
                 ],
             ])->willReturn($builder)